[Refactor] 병합 후 kafka 수정
